F01 Project clean up
