+++ Top100Albums/Top100Albums/Models/Genre.swift
@@ -1,0 +1,14 @@
+//
+//  Genre.swift
+//  Top100Albums
+//
+//  Created by Rolando Rodriguez on 1/18/21.
+//
+
+import Foundation
+
+struct Genre: Codable {
+    let id: String
+    let name: String
+    let url: URL
+}
